dev: create controller common
