+++ packages/api/src/common/controller.ts
@@ -1,0 +1,12 @@
+import { ApiModelProperty } from '@nestjs/swagger'
+import { IsNumberString } from 'class-validator'
+
+export class ParamsWithId {
+  @ApiModelProperty()
+  @IsNumberString()
+  readonly id: number
+
+  constructor (id: number) {
+    this.id = id
+  }
+}
