+++ packages/core/src/page/page.service.ts
@@ -1,5 +1,5 @@
 import { ApiModelProperty, ApiModelPropertyOptional } from '@nestjs/swagger'
-import { IsString, IsNumberString, IsUrl, IsOptional } from 'class-validator'
+import { IsString, IsUrl, IsOptional } from 'class-validator'
 import { Repository, InjectRepository } from '../common/service'
 import { Optional } from '../common/optional'
 import { Page } from './page.entity'
@@ -31,16 +31,6 @@ export class UpdatePageData {
   readonly url?: string
 }
 
-export class ParamsWithId {
-  @ApiModelProperty()
-  @IsNumberString()
-  readonly id: number
-
-  constructor (id: number) {
-    this.id = id
-  }
-}
-
 export class PageService {
   constructor (
     @InjectRepository(Page)
@@ -51,7 +41,7 @@ export class PageService {
     return this.pageRepository.find()
   }
 
-  async get ({ id }: ParamsWithId): Promise<Optional<Page>> {
+  async get (id: number): Promise<Optional<Page>> {
     const page = await this.pageRepository.findOne({ id })
     return Optional.ofNullable(page)
   }
@@ -60,7 +50,7 @@ export class PageService {
     return this.pageRepository.save(new Page(name, url))
   }
 
-  async update (id: ParamsWithId, data: UpdatePageData): Promise<Optional<Page>> {
+  async update (id: number, data: UpdatePageData): Promise<Optional<Page>> {
     const optionalPage = await this.get(id)
     if (optionalPage.isPresent()) {
       const page = await this.pageRepository.save({ ...optionalPage.get(), ...data })
@@ -69,7 +59,7 @@ export class PageService {
     return optionalPage
   }
 
-  async remove (id: ParamsWithId): Promise<Optional<Page>> {
+  async remove (id: number): Promise<Optional<Page>> {
     const optionalPage = await this.get(id)
     if (optionalPage.isPresent()) {
       await this.pageRepository.remove(optionalPage.get())
