+++ packages/core/src/typeorm.config.ts
@@ -3,11 +3,11 @@ import { TypeOrmModuleOptions } from '@nestjs/typeorm'
 
 const TypeOrmConfig: TypeOrmModuleOptions = {
   type: 'postgres',
-  host: '58.181.37.175',
-  port: 60016,
-  database: 'test',
-  username: 'sodas_dev_user',
-  password: 'sodas_dev_secret',
+  host: 'localhost',
+  port: 5432,
+  database: 'webok_dev',
+  username: 'webok_dev_user',
+  password: 'webok_dev_secret',
   entities: [path.join(__dirname, '**/*.entity{.ts,.js}')],
   // Migrations
   migrationsTableName: 'typeorm_migrations',
